Move InsightsUpsellTab to metabase/common/components/upsell
