+++ frontend/src/metabase/common/components/upsells/InsightsUpsellTab/InsightsUpsellTab.tsx
