+++ frontend/src/metabase/common/components/upsells/InsightsUpsellTab/index.ts
@@ -1,0 +1,1 @@
+export * from "./InsightsUpsellTab";
